# release-2.1: opt: Improve performance of the cost-based optimizer's PREPARE phase